+++ ui/src/main/java/io/snabble/sdk/ui/cart/shoppingcart/adapter/widgets/PriceDescription.kt
@@ -1,0 +1,34 @@
+package io.snabble.sdk.ui.cart.shoppingcart.adapter.widgets
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import io.snabble.sdk.ui.cart.shoppingcart.row.ProductRow
+
+@Composable
+fun PriceDescription(row: ProductRow) {
+    Column(
+        horizontalAlignment = Alignment.Start,
+        verticalArrangement = Arrangement.spacedBy(4.dp, Alignment.CenterVertically)
+    ) {
+        row.name?.let {
+            Text(
+                text = it,
+                overflow = TextOverflow.Ellipsis,
+                maxLines = 2,
+                style = MaterialTheme.typography.bodyLarge
+            )
+        }
+        row.priceText?.let {
+            Text(
+                text = it,
+                style = MaterialTheme.typography.bodySmall
+            )
+        }
+    }
+}
